Die gewünschte Zielrichtung kann jetzt auch mit 'n', 'o', 's' und 'w' 	angegeben werden.

+++ SPEC_A_2/S_Drehen.h
@@ -33,7 +33,7 @@ class S_DrehenClass : public StateClass
 	 void Think();
 	 void Act();
 
-	 int ZielRichtung;
+	 byte ZielRichtung;
 	 
 };
